feat(lib): guard the API graph fs-pure and split git resolution out of local_path_policy
- src/api/fs_purity_test.ts walks the runtime (code) module graph of
  src/api/mod.ts via deno info and fails on any subprocess or network
  API in first-party modules; statically-visible dynamic imports are the
  sanctioned seam for CLI-only capability
- repository-source git resolution moves to repository_source_git.ts,
  reached through repository_source.ts, which loads it lazily at the
  moment a repository/floating source is resolved; versionPayloads
  refuses such sources up front, so the library graph stays
  subprocess-free under static analysis
- tryRunGit feature-detects Deno.Command via globalThis so the dnt/Node
  build neither type-checks nor links against a subprocess API its shim
  does not provide; without it, git resolution degrades to no-match
  instead of crashing
