+++ src/runtime/artifact_resolution/resolver.ts
@@ -17,8 +17,8 @@ import {
   LocalPathAccessError,
   type OperationalLocalPathPolicy,
   resolveAllowedLocalPath,
-  resolveRepositorySourceFloatingLocalPath,
 } from "../operational/local_path_policy.ts";
+import { resolveRepositorySourceFloatingLocalPath } from "../operational/repository_source.ts";
 import type {
   ArtifactResolutionContentMode,
   ArtifactResolutionContext,
